주변장소 맵 카테고리 검색 범위 설정 오류수정 완료

+++ app/src/main/java/com/zerodsoft/scheduleweather/activity/preferences/fragments/SettingsFragment.java
@@ -185,7 +185,7 @@ public class SettingsFragment extends PreferenceFragmentCompat implements Prefer
                             @Override
                             public void onClick(DialogInterface dialogInterface, int i)
                             {
-                                String value = String.valueOf(slider.getValue() * 1000);
+                                String value = String.valueOf((int) (slider.getValue() * 1000));
                                 searchMapCategoryRangeRadiusPreference.callChangeListener(value);
                                 dialogInterface.dismiss();
                             }
